+++ Changelog.md
@@ -25,7 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bug fix sqlite engine cannot save log and db to specific folder. Completed on 2024-07-02
 - Bug fix mariadb engine register failure not throw error. Completed on 2024-07-02
-- Bug fix in `slqtemplate.js` which unable create second key value during generate sql update statement . Completed on 2024-07-02
+- Bug fix in `slqtemplate.js` which unable create second key value during generate sql update statement. Completed on 2024-07-02
+- Bug fix in `query` function at `mariadb.js` to check the sql statement must in string type before proceed the data query to predict mariadb module crash. Completed on 2024-07-03
 
 ### Security